no longer support local and standalone cli modes
