+++ bin/ksql-stop
