+++ ksql-cli/src/main/java/io/confluent/ksql/util/CliUtils.java
@@ -33,8 +33,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Map;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import io.confluent.ksql.ddl.DdlConfig;
 import io.confluent.ksql.parser.AstBuilder;
@@ -86,7 +84,7 @@ public class CliUtils {
     }
   }
 
-  public String readQueryFile(final String queryFilePath) throws IOException {
+  public String readQueryFile(final String queryFilePath) {
     StringBuilder sb = new StringBuilder();
     try (final BufferedReader br = new BufferedReader(new InputStreamReader(
         new FileInputStream(queryFilePath), StandardCharsets.UTF_8))) {
@@ -114,28 +112,6 @@ public class CliUtils {
       );
     }
     return new PropertiesList(propertiesList.getStatementText(), properties);
-  }
-
-  private static final Pattern QUOTED_PROMPT_PATTERN = Pattern.compile("'(''|[^'])*'");
-
-  private String parsePromptString(String commandStrippedLine) {
-    if (commandStrippedLine.trim().isEmpty()) {
-      throw new RuntimeException("Prompt command must be followed by a new prompt to use");
-    }
-
-    String trimmedLine = commandStrippedLine.trim().replace("%", "%%");
-    if (trimmedLine.contains("'")) {
-      Matcher quotedPromptMatcher = QUOTED_PROMPT_PATTERN.matcher(trimmedLine);
-      if (quotedPromptMatcher.matches()) {
-        return trimmedLine.substring(1, trimmedLine.length() - 1).replace("''", "'");
-      } else {
-        throw new RuntimeException(
-            "Failed to parse prompt string. All non-enclosing single quotes must be doubled."
-        );
-      }
-    } else {
-      return trimmedLine;
-    }
   }
 
   public static String getLocalServerAddress(int portNumber) {
