+++ ksql-cli/src/test/java/io/confluent/ksql/cli/RemoteCliSpecificCommandTest.java
@@ -37,7 +37,7 @@ public class RemoteCliSpecificCommandTest {
 
   private final StringWriter out = new StringWriter();
   private final KsqlRestClient restClient = new KsqlRestClient("xxxx", Collections.emptyMap());
-  private final RemoteCli.RemoteCliSpecificCommand command = new RemoteCli.RemoteCliSpecificCommand(restClient, new PrintWriter(out));
+  private final Cli.RemoteServerSpecificCommand command = new Cli.RemoteServerSpecificCommand(restClient, new PrintWriter(out));
 
   @Test
   public void shouldRestClientServerAddressWhenNonEmptyStringArg() {
@@ -61,7 +61,7 @@ public class RemoteCliSpecificCommandTest {
 
   @Test
   public void shouldPrintErrorOnErrorResponseFromRestClient() {
-    final RemoteCli.RemoteCliSpecificCommand command = new RemoteCli.RemoteCliSpecificCommand(
+    final Cli.RemoteServerSpecificCommand command = new Cli.RemoteServerSpecificCommand(
         new KsqlRestClient("xxxx", Collections.emptyMap()) {
           @Override
           public RestResponse<ServerInfo> getServerInfo() {
